Fix bug on 64-bit machines
git-svn-id: svn://svn.code.sf.net/p/libsh/code/sh/trunk@2951 afdca40c-03d6-0310-8ede-e9f093b21075

+++ backends/cc/CcEmit.cpp
@@ -85,7 +85,7 @@ typedef std::map<SH::ShOperation, CcOpCodeVecs> CcOpCodeMap;
 CcOpCodeVecs::CcOpCodeVecs(const CcOpCode &op) {
   std::string code = op.code; 
 
-  unsigned i, j;
+  size_t i, j;
   i = j = 0;
   for(; (j = code.find_first_of("#$", i)) != std::string::npos;) {
     frag.push_back(code.substr(i, j - i));
